package manager: declarative dependency calculation.
This is actually getting there. At this point, it can calculate
which versions of packages are needed to satisfy a set of user-
defined requirements of the form

    package v1 v2...

This is the same format that is used in the METADATA repo to give
the dependencies of each package version (yes, this is per-version).
The first item is the package name, followed by any number of
version numbers, in ascending order. These a series of intervals of
versions. Specifying no version numbers means that any version will
do. Giving only one sets a lower bound; two gives a lower and upper
bound; three gives one interval and then a lower bound, etc.

The solve function uses linear programming (twice!) to figure out
what the optimal set of package versions is to satisfy the given
requirements. Newer and fewer packages are preferred, with weight
determined (also by linear programming) to favor versions that
satisfy more dependencies and are newer. Tweaking exactly what
gets installed can be accomplished by adding requirements.

This declarative approach to managing what's installed needs to be
tied together with a set of commands that can manipulate it easily,
but at this point, the core functionality is there.