Make $PATH changes permanent in Linux (Ubuntu)
Maybe this isn't true for other distributions (I'm no Linux guru), but in Ubuntu the only way to make the export command's changes to $PATH stick is to add it to .bashrc so that it is executed every time a shell is started. Since Ubuntu is the distribution with more users, it makes sense to at least mention this method.