(re-)fix bash export command in Binary Installation instructions
In my last pull request I proposed to "harmonize" the two bash export commands that were used in this file, but actually the context of the second one is different (it isn't suggested to cd into the julia directory before, while in the first example it is). This commit restores the correct command.