Fix complex matrix multiply when both inputs are refs to the same matrix
The conj! trick, which I introduced earlier, doesn't work when both
inputs are refs to the same matrix. I fixed this by calculating the
proper value on each access. To make this fast, I introduced the
previously-proposed LapackFlag, so it can be done by dispatch. It would
be good to propagate this to the rest of the Lapack suite, but that
will take time I don't have at the moment.