Update randmtzig.c to use dsfmt's double precision RNG.
Earlier, randmtzig used the int32 RNG from dsfmt, which is
not the preferred way to use dsfmt. 52-bits of randomness
are now used from the mantissa of the doubles generated by
dsfmt. This is two fewer bits than the original code, but
much faster.

This passes the Testu01 crush testsuite, when the output of
randn() is mapped through the inverse CDF into the uniform
distribution. The testsuite is at
https://github.com/andreasnoackjensen/rngtest.jl

Detailed discussion in https://github.com/JuliaLang/julia/pull/1372
Close #1348.