isequal(::Set, ::Set) + tests.
The current behaviour when comparing sets of unrelated types is not
particularly consistent. (I suspect that similar inconsistencies will
be present in the implementations of intersect, setdiff and union.)
The tests describing this behaviour do not necessarily state that this
is the correct behaviour, thery merely document what is currently
present.

The implementation was provided earlier, mixed in with the tests for
Set. Now the implementation has been moved to base/set.jl, and some
tests for it have been added to test/corelib.jl

A helper macro (@assert_raises) was added (in test/corelib.jl), to
make it easier to express tests about error conditions. This should be
moved out to some more sensible place.