extras: remove ones that have been made into packages
The rest of the things in extras need to be made to use modules
and turned into packages also. The initial batch were the ones
that were already using the module system, so it was clear what
to do with them. I also skipped anything requiring wrappers as
I'm not certain how that's going to work yet.