Pkg: don't auto-commit after adds and rms; this was annoying
Additions and removals are now automatically staged but no commit
is done. If you want to commit a .julia repo state, you have to do
it yourself. Many users may simply never do this at all, which is
the minimally invasive behavior. If you want git version history
for your packages, it's there for you, but you needn't use it.