Test: get rid of broken @assert_raises macro, use @test_fails.
The @assert_raises macro was quite broken: it would pass even if
the tested expression did NOT raise the expected exception type.
The only way it would actually raise an error was if an exception
that was not of the indicated type was raised. Comparison of non-
similarly typed Dicts seems to work now, but the assertion looked
like it should be failing.