Time: simple Time type with partial NumPy datetime64 compatibility
Relevant discussion:

https://groups.google.com/d/topic/julia-dev/YlUS7899gro/discussion

This is mostly a straw man, and not meant to be definitive, but I
figure it gets the ball rolling, and maybe complementary to Mike's
Calendar package [1]. The idea would be that there are really two
very different ways of keeping track of time: SI seconds and the
powers thereof, which has nothing to do with calendar time (days,
weeks, months, years), and calendar time, which has everything to
do with that. The tricky part is converting between them, but for
many, many problems, one really wants to work entirely in one of
the two domains, rarely, if ever crossing into the other.

[1] https://github.com/nolta/Calendar.jl