Char: remove many (unnecessary) Char arithmetic operations.
Spurred by #1771. The one I really wanted to get rid of was
Char+Char but that's used by iteration of Char ranges, e.g. here:

https://github.com/JuliaLang/julia/blob/63dd47153097c68643dd3462656a6c74358d04e3/base/intfuncs.jl#L220

In order to get rid of that usage, the way Range iteration works
would need to be changed. Currently it's a bit odd that 'a':'z'
is a Range of Chars, while 'a':1:'z' is a Range of Ints. Maybe
the step type of a Range shouldn't need to be the same as the
start type?