improve behavior of constants
the old behavior was simply bad because it allowed a program to continue
without executing some assignments.
now we allow changes to constants that won't break things, namely those
where the new and old values are egal, or are simply of the same type but
not themselves types or functions.
a warning is still given, to indicate that this is risky and might change.
this feature is mostly for interactive development, and there is a notion
that production code should not print warnings.
any other changes to constants are errors.
fixes #1389, fixes #1756