stitch Base together with importall instead of using.
This addresses an issue discovered by @johnmyleswhite: if you did
importall Base and then tried to add methods to exp2, for example,
or anything else defined in Math or any other submodule of Base,
because of the fact that Base was only *using* that name but then
exported it, modification failed. If you importall something and
then export it, however, another module can import that symbol and
modify it. It should probably be an error or a warning to export
something that you only have access to via using.