JL_ANSWER_COLOR => JULIA_ANSWER_COLOR; JL_POST_BOOT gone [close #2076]
JL_PRIVATE_LIBDIR remains and I wasn't sure what to do about that. It
seems to be unused at this point -- or at least I couldn't find any
usages of it.

This commit also changes the way one controls the answer color during
a repl session: there is now a Main.ANSWER_COLOR variable that you can
modify to change the answer color on the fly. The environment variable
JULIA_ANSWER_COLOR is only checked upon repl startup and not after.