Regex: remove execution-time regex options [closes #2077].
This deletes the following methods:

    - ismatch(re,str,opts)
    - match(re,str,idx,opts)

It is kind of a weird idea that you can modify the meaning of a regex
by passing options at match time. Instead, you should create the regex
with those options and the matching functionality decides when to pass
those options to PCRE -- at compile or exectute time.

At some point in the future, ismatch(re,str,idx) should be added, but
at the moment, since this is the exact same signature as the deleted
method, we should wait and let people's code break instead. Although
frankly, I really doubt anyone has been passing execution options to
PCRE.