improvements: convert(Unsigned,x), flipsign, and indirectly abs().
This changes conversion to unsigned so that for any signed value x
signed(unsigned(x)) == x, even when x is negative. In order not to
break things, this required changing flipsign on 64-bit systems so
that it promotes to Int64 first, which also has the effect that
abs(Int32) produces an Int64 with the correct value. Previously it
always produced and Int32 and for typemin(Int32).