sort(f,v) sortby(f,v) and export {Insertion,Merge,Quick,Tim}Sort.
This allows most common use cases of the sorting API to avoid having
to reach for stuff inside of the Sort module. If you need apply the By
ordering to select or sortperm you still may need it. Exporting the
sort algorithm types is not likely to clash with anything since they
all end with "Sort".