replace {key,val}type for associatives with eltype returning (K,V).
We don't actually use these two functions anywhere in Base, so it's
hard to say it this makes things more awkward or not, but I suspect
that a typical usage actually wants both the key and value types, so
you can just write this:

     K, V = eltype(a)

That's more convenient than having to call both keytype and valtype.
In any case, getting rid of two exports is a big win.