Bait someone into fixing dev/release branch version testing
Because it's presumably too late to get this into Ubuntu's
release-0.1, package authors should define the following function:

function branch()
  local jbranch
  try
      jbranch = Base.branch()
  catch
      jbranch = VERSION
  end
  jbranch
end

Alternatively, bumping the VERSION file to 0.2.0-pre (or whatever
syntax it needs) seems like a better approach. But currently both
release-0.1 and master claim to be 0.1.0, which seemingly makes it
impossible to know what branch you're on.