Pkg: use "devel" branch of METADATA.jl for future Julia versions.
The default "master" branch is going to be used to maintain a fork of
METADATA.jl that doesn't contain any package versions that require
newer Julia versions (i.e. 0.2.x). The older METADATA fork will also
no "julia" dependcy lines. There is a "./remaster.jl" script in the
METADATA.jl repo that is used to maintain the "master" branch from
the "devel" branch automatically (needs to be run manually though).