Try to find modules in PACKAGE.jl as well
This change adds a case to `find_in_path` that attempts to find
directories appended with ".jl" as well.

The trend seems to be naming package repositories as PACKAGE.jl, but you
then had to clone packages to a local directory stripped of the ".jl" so
that files could be found.

No more!