pkg2: add .julia directory finding and creation.
I'm not sure that we still want .julia to always be a git repo now,
but for the time being, it's easiest to just steal code from Pkg 1.
In any case, the business of prompting users for their GitHub info
is a complete annoyance and has to go. We should probably get rid
or that in the current Pkg, but I'm loathe to change anything there.