loading: . in LOAD_PATH redundant; always look in Pkg.dir() first.
This closes #2968 since changes to JULIA_PKGDIR are immediately
reflected in Pkg.dir() which is now always considered when requiring
things before looking in LOAD_PATH. Having "." in LOAD_PATH was
redundant since the find_in_path logic *already* looks for files
that exist first. I suspect that all of this logic needs some
regularization because it's pretty lumpy right now, but this does
the trick and closes the issue without breaking things.