statistics: use select! for median!; export median! and quantile!
Using select! should make median computations much faster since we
only partially sort the input array. It would be nice to avoid the
second call to select! in the case of an even-length array, but it
isn't obvious how to do this. Perhaps we need a generalized select
that selects the values at a range of indices, in sorted order.