sysimg.jl: use relative includes during later system image building.
Before this, it was impossible to both load Pkg2 inside sysimg.jl and
load it at the repl for testing, since the two contexts had different
understandings of what it means to include a relative path. With this
change, the later stages of sysimg building use the same kind of
relative include semantics as the normal (i.e. repl) include.

This change also affects linalg/{cholmod,umfpack}.jl, making them
potentially possible to load from the repl, not just from sysimg.jl.