pkg2/resolve: introduce VersionWeigths
This changes resolve's objective function such that
version pruning doesn't affect it. Therefore, equivalence
classes are no longer needed in resolve (they're still used
in sanity_check).

With this change, the objective function is defined such that:

1) requirements and dependencies are enforced
2) in case of multiple solutions, required packages versions
   are maximized:
   2a) the overall sum of required packages major versions is
       maximized
   2b) in case of a tie, minor versions are considered; if there
       are still ties, patch versions are considered
3) in case of multiple solutions under the above rules, non-required
   package versions are maximized, still counting major versions
   first, then minor, then patch, as for the required versions.
4) if there is still a tie between different solutions, dependants
   are maximized before dependencies
5) if there is still a tie, random noise is used to pick one

The change introduced by this commit is the major/minor/patch
count. Also, package installation is now considered "cheap", while before
the cost of installing a package was equivalent to that of downgrading
a package.