Pkg2: checkout(pkg,[branch]) and release(pkg).
These are the names I settled on for checking out a branch version
of a package and then releasing the package back to management. I
started out having a `recur` option, but that turns out to be kind
of a nightmare to make work and is probably not needed that often.