big Pkg rename: Pkg, Pkg2 => Pkg1, Pkg; sysimg loads only Pkg(2).
Most of the functionality of Pkg1 has now be replicated in Pkg2 and
seems to be far more stable and usable than Pkg1 at this point, so
this renames them and makes the system image only include the new
Pkg(2). You can still use Pkg1 by explicitly loading it:

    include(joinpath(JULIA_HOME,"../../base/pkg1.jl"))

Hopefully all of Pkg1's functionality will be subsumed shortly.