randmatmul implemented by calling out to libblas.
This brings the JS benchmark close to the others which
also mostly call out to libblas. Note that about 90% of the time
for the rand_mat_mul benchmark is now taken up by random matrix
generation.

IMPORTANT: This change makes the Javascript benchmark run only in
node.js. The 'ffi' and 'ref' npm modules need to be installed. libblas
must also be available as a dynamically loadable library. The npm modules
are installed as -

    npm install ffi
    npm install ref