`using Nonexistent`: better behavior loading unfound names.
    julia> using FooBar
    ERROR: FooBar not found
     in require at loading.jl:45

The core change is that `find_in_path(name)` now returns `nothing`
when `name` isn't found and various functions that call it must
check explicitly for `nothing` as a return value and throw an error
as appropriate. In particular `require` and `reload` now raise an
error instead of just forging ahead trying to load a file we know
doesn't exist, which resulted in less-than-clear error messages.