Pkg.scaffold: generate .travis.yml and setup GitHub URL by default.
This won't happen unless you've got `git config github.user` setup
but if that is the case, this automatically generates a minimal file
that will test that the module that the package is expected to provide
loads correctly. It also automatically includes the build status icon
in the README.md file (unless you opt out).