Added footnote in parallel computing chapter on MPI
The Julia manual makes the statement that Julia's parallel programming model is
unlike MPI. This is not entirely true. The current MPI standard includes
mechanisms for remote memory access; i.e. one-sided communication. Thus, users
familiar with the latest MPI standard may find the documentation confusing.

To rectify this problem, a footnote was added to clarify the meaning of the term
"MPI", as used by the Julia manual.