Export watch_file.
Export watch_file, as this is the easiest API entry point between FileMonitor
and PollingFileMonitor.

Fix _uv_hook_fseventscb such that the condition is notified even if no callback
is in the monitor. Also fix the value sent to notify() to match the expected
return value in wait().

Fix watch_poll default callback.