Pkg: factor out META_BRANCH (currently "devel").
This paves the way for METADATA versioning via branches. The idea
is that each version of the METADATA format has its own branch.
Incompatible formats can co-exist peacefully because they each have
their own branch in the METADATA repo. Some sort of automated version
of the `remaster.jl` script would be ideal for pack-porting packages
that can be backported to earlier versions. This should be possible
with GitHub service hooks.