eliminate some spurious method matches during type inference. fixes #4374
these occurred when a method matched, but the intersection was actually
already covered completely by prior (more specific) matches. the relevant
case here was methods(+,(Int,Int...)), which matches (Int,Int) and
(Number,Number), with the second one obviously redundant. the tricky part
is that the (Int,Int) match does not fully cover (Int,Int...), which
is what causes us to look for more matches.

this change cooperates with the prior improvements to inlining of vararg
functions to remove all overhead from + with any number of arguments.