warn: remove `depth` keyword, no backtraces; warn(err::Exception).
Since it's completely unclear what this does or how to use it, I'm
deleting this feature. When `depth=0` backtraces seem to show, while
when `depth=1` they don't or something like that. Now we just don't
ever show backtraces for warnings. This keyword didn't do anything
for the info function since we never print backtraces there.

Also added a warn(err::Exception) method that prints an error the
way it would look if allowed to kill the current task, but doesn't
actually kill the current task. This is useful if catch an error
and want to show it to the user but continue running.