Further changes to eigs
(1) ARPACK info flag 1 now results in a warning (maximal number of
iterations taken) rather then throwing an error.
(2) Added the return value nconv, indicating the number of converged
eigenvalues, which is relevant when the maximal number of iterations
has been taken before all requested eigenvalues have converged (which
no longer leads to an error thanks to (1) ).
(3) Modified behavior of eigs such that any input type of matrix is
allowed, as long as the type of A overloads the methods size, issym,
eltype and can be multiplied with a vector of the same element type.
(4) Updated test file accordingly