allow Pkg.clone(pkg) to work [close #4544].
If you do Pkg.clone(pkg) where pkg is a registered package, it is now
the same as doing Pkg.clone(url,pkg) where the url is the registered
url for the package. It's not quite the same as Pkg.add(pkg) followed
by Pkg.checkout(pkg) since it doesn't add pkg to .julia/REQUIRE. This
means that if you do Pkg.release(pkg) after Pkg.clone(pkg) it will
simply remove the package altogether. This may not be the desired
behavior ultimately, but it's what it does right now.