Enble LLVM support for Intel VTune Amplifier if platform supports it.
The change to deps/Makefile causes LLVM to be compiled with --with-intel-jitevents for
* Windows
* Linux on x86

The change to src/codegen.cpp causes Julia to register a IntelJITEventListener
when the environment variable ENABLE_JITPROFILING is set.

Intel VTune Amplifier is not required to compile the code because the standard LLVM
distribution (3.3 or current pre-3.4 development version) already has the necessary code.