base/test.jl: fix the broken {register,with}handler API; export less.
This was never meant to work like that. There is a stack of handlers,
not an array slot that you assign to. And the stack must be per-task,
so using a global array is broken too. This version uses task-local
storage instead to do dynamically scoped test result handlers.

I'm not sure why so many things were exported from the Test module.
Things that are prefixed with test_ and are clearly meant to be used
from other modules make sense to export but exporting names like
Result, Success, Failure, Error, registerhandler, and withhandler is
just nuts – those should be accessed with explicit qualification as
being inside the Test module.