test/bigint.jl: adjust so bigfib(0) == 0, add more tests.
I noticed while doing this that this uses a non-standard definition
where fib(0) == 1 instead of fib(0) == 0. Doesn't really matter as
this is just a test, but it caused me some concern when this result
wasn't checking out with fib(10^6) as computed by Wolfram|Alpha™.
So I changed this to match the standard definition and shortened it
up a bit. I also added a bunch of tests to make sure that it works
as it should.