Added collect with eltype as argument
Currently collect(itr) does not allow the user to specify the type of
the elements of the returned array, but relies on the eltype() function.
This commit adds collect{T}(element_type::Type, itr::T), so that a user
can get a type specific array, without having to define eltype on the
iteration object.

Inspiration for this julia-users thred:
https://groups.google.com/forum/#!topic/julia-users/UbwDusuZa1o