mod2pi: remove modpi and modpio2, adding a single function to Base.
This is kind of a painful edit since there's a lot of great work in
here, but it's just too much to add three functions to Base. mod2pi
seems like the most sensible of these functions. If we find that
there is great need for the other versions, we can always resurrect
them in the future by the wonders of git history.