move Git into Pkg – I probably should have done this long ago.
This is acknowledges that the Git module is not really a separate
or external API, it's really just a supporting module for Pkg. This
has the important effect that you can reload Pkg – including Git –
just by doing `include(joinpath(JULIA_HOME,"../../base/pkg.jl"))`.
Previously, you could reload Pkg that way, but not Git.