fix macroexpansion bug in keyword arguments
The names of keyword arguments are externally visible, so they are very
hard to rename. This commit simply avoids renaming or qualifying them,
so collisions are possible but without them everything works fine (which
was not the case before). To do this better we'll have to rename them but
also carry around the original symbol for matching purposes, which will
be pretty annoying.