parseint: speed up parsing by avoiding checks in main paring loop.
The trick here is that we know that we can't possibly overflow until
we reach a fairly large value for each type. In particular, we can't
possibly overflow while n <= div(typemax(T)-base+1,base). After that
we can have one or two decoding cycles left, which actually have to
be checked for overflow. The "naïve" version was 2x+ slower than the
unchecked version, while this version is only around 15% slower.