Pkg.generate: use `git shortlog` to get authors for existing repos.
This is trickier than it should be because `git shortlog` is very
broken: it refuses to print any output unless it is connected to a
TTY. I've tried to make this work on Windows, but have no way of
testing it. According to the Internets, "CON:" is the equivalent of
"/dev/tty" on Windows.