fix #5084; in base/range.jl, minimum, maximum: use ?: not &&, ||.
A slightly sketchy use of && and || as control flow got into range.jl
with unfortunate consequences for performance. Reminds me of just how
delicate that file is. I suspect that we *could* enhance the inference
for && and || in this case, but I'm not positive. In any case, using
a "real" control-flow construct like the ternary operator works too.