Profile.print(format=:flat) – sort in descending order of samples.
This seems like the most useful order for immediately seeing what
the hotspots are. Doing this with the text output is much harder
than doing it with the raw data. It would be nice to return this
data in a form that you can easily manipulate rather than just
printing, but that's an issue for another day.