fix #5401
type inference was being too greedy by returning `None` for no matching
methods. it's safer to be able to handle methods that might not be
defined yet. we might be able to roll back this change when #265 is
fixed.

also fix a bug in ... with keyword arguments that caused https://github.com/dcjones/Gadfly.jl/issues/157