Changed Dict(ks,vs) to behave like Dict(zip(ks,vs))
Currently the lenght of ks is used to deterimine the numer of elements
in the dictionary. After this, the shortest of ks and vs is used.

An alternaltive might be to throw an error if they are of equal lenght,
but I think the same behaviour as Dict(zip(ks,vs)), is the most
intuitive.