windows: add custom SFX file for installer
This adds a custom self-extracting binary for the Julia Windows installer.
As discussed in b013e9bf53, the installer package has several parts:

[7zS.sfx | config.txt | 7z payload ]

7zS.sfx extracts the payload and then runs the installer executable specified
in `config.txt`.

The change here is to use "asInvoker" permission request in the manifest.
Recent Windows versions have a broad pattern match to determine whether an exe
is an "installer". Any file that matches must be run with Admin rights, unless
the necessary level is specified in the application manifest. We set the level
in the NSIS config as well, but it needs to be set for 7zS.sfx because the
NSIS-generated installer is subordinate.

Additionally the Julia icon is now used for the installer package.

The customization is based on the 7zS.sfx file in 7z920_extra.7z from 7-zip.org
7zS.sfx is under LGPL, so this modified version is as well.

Modification steps:

1. download Resource Hacker from here: http://www.angusj.com/resourcehacker/
2. open 7zS.sfx in Resource Hacker

4. Action->Add a New Resource
   Click <Open file with new resource>
   Select "contrib/windows/7zS.sfx.manifest"
   Under Resource [Type,Name,Language] enter:

       24, 1, 1301

   These are: RT_MANIFEST, [Name=1], LANG_ENGLISH
   Per documentation:
   http://msdn.microsoft.com/en-us/library/ms648009.aspx
   http://msdn.microsoft.com/en-us/library/dd318693.aspx

5. Action->Replace Icon
6. <Open file with new icon>
   Select contrib/windows/julia.ico
   Various icon zoom levels will be created automatically
7. Save 7zS.sfx