Make f, d, def actual let variables in test/collections.jl (fixes #5470)
When test/core.jl, test/collections.jl, and test/remote.jl are all
run on the same core, a global variable :f is defined in test/core.jl
in the test for issue #2619, which was then redefined inside of a let
block by the (recently added) `get!` test in test/collections.jl,
which was propagated to test/remote.jl and caused #5470.

This commit makes f an actual let variable, after which f does not
seem to be global in test/remote.jl, and the local def works.