Fix #3567
We malloc an extra byte at the beginning of the uv objects that have buffers (and thus might get a callback
that the buffer was filled even when not blocked on a read). It might be worth putting it at the end instead,
but that's kind painful too, since they are all of different length.