Allow tripple qouted strings as format strings for `[@]printf`
This is a temporary hack for #2682, but the proper fix would be to fix
the parser to not emit macroall to the AST for triple quoted strings.

This commit also includes a ugly hack that will error when it is time to
remove this fast fix.

See: https://groups.google.com/forum/#!topic/julia-users/U1shoGwnCCQ