Fix two small bugs in process handling with new libuv.
First, I forgot the finalizer (my bad). The second issue is slightly
more subtle which is that we now destroy process objects earlier, so
libuv might actually get a SIGCHLD. We have a custom patch that
disables signals while in fork, which also happended to block SIGCHLD
because it was in the wrong place. The associated libuv commit fixes
that.