mitigate #6097
the general problem was that one place in the code had a larger type for
a tuple than another. I don't know why, but we can at least generate
correct code instead of crashing. there was already some code in boxed()
towards handling this kind of thing, but it needed to be rearranged a bit.