Move -D_FILE_OFFSET_BITS back to CFLAGS
These should really go in CPPFLAGS, but apparently on mingw, there's defines
in the stdio headers that rename the function (rather than doing it on a symbol lelve), so we can't use this when we include LLVM headers. Anyway, as long as
we don't call the IO functions from C++ we should be fine.